+++ src/components/stages/header/Search.jsx
@@ -1,14 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import debounce from '../../../helpers/debounce';
 
 const Search = ({ activeNavigation, searchButtonLabel, searchInformationItems, searchTextPlaceholder }) => {
-    const search = () => {
+    const search = debounce(() => {
         searchInformationItems(activeNavigation, document.getElementById('searchPattern').value);
-    };
+    }, 200);
 
     return <div className={'searchContainer'}>
         <input id={'searchPattern'} type={'text'} className={'searchContainer__inputField'}
-               placeholder={searchTextPlaceholder} onKeyUp={() => search()} />
+               placeholder={searchTextPlaceholder} onKeyUp={() => search() } />
         <input type={'button'} className={'button'} className={'searchContainer__button'}
                value={searchButtonLabel} onClick={() => search()} />
     </div>;