+++ CHANGELOG.md
@@ -5,9 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 2020-10-30
+- Changed | item wishlist-icon is now a separate container/component
+- Changed | item link bar now a separate container/component
+- Added   | item link bar to wishlist
+
 ### 2020-10-29
 - Added   | provider links in overview
 - Changed | wishlist-icon is now bound to item image
+- Changed | debounced search field input
 
 ### 2020-10-27
 - Added   | navigationId is now set from the address-bar path if the page is refreshed